math commands

+++ src/command/math/Difference.java
@@ -15,7 +15,7 @@ import command.Command;
 public class Difference extends Command {
 
 	@Override
-	public int execute(List<Integer> argument) {
+	public double execute(List<Double> argument) {
 
 		return argument.get(0) - argument.get(1);
 	}
